backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.54.0
chromadb==0.4.22
sentence-transformers==2.3.1
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.0
lxml==5.1.0
